backend/requirements.txt
fastapi==0.115.0
pydantic-settings==2.4.0
uvicorn[standard]==0.30.6
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
supabase==2.7.4
python-dotenv==1.0.1
aiofiles==23.2.1
Pillow==11.2.1
